+++ R/simdata.R
@@ -4,7 +4,8 @@
 #' from multivariate normally distributed factor (\eqn{\theta}) scores, or from
 #' a user input matrix of \eqn{\theta}'s.
 #'
-#' Returns a data matrix simulated from the parameters.
+#' Returns a data matrix simulated from the parameters, or a list containing the data,
+#' item objects, and Theta matrix.
 #'
 #' @param a a matrix of slope parameters. If slopes are to be constrained to
 #'   zero then use \code{NA}. \code{a} may also be a similar matrix specifying
@@ -38,6 +39,9 @@
 #'   of zeros
 #' @param Theta a user specified matrix of the underlying ability parameters,
 #'   where \code{nrow(Theta) == N} and \code{ncol(Theta) == ncol(a)}
+#' @param returnList logical; return a list containing the data, item objects defined 
+#'   by \code{mirt} containing the population parameters and item structure, and the 
+#'   latent trait matrix \code{Theta}? Default is FALSE
 #' @author Phil Chalmers \email{rphilip.chalmers@@gmail.com}
 #' @references 
 #' Reckase, M. D. (2009). \emph{Multidimensional Item Response Theory}. New York: Springer.
@@ -172,17 +176,21 @@
 #' #mod <- mirt(dataset, 1, itemtype = c('2PL', '2PL', '2PL', '2PLNRM'), key=c(NA,NA,NA,1))
 #' #coef(mod)
 #' #itemplot(mod,4)
+#' 
+#' #return list of simulation parameters
+#' listobj <- simdata(a,d,2000,items,nominal=nominal, returnList=TRUE)
+#' str(listobj)
 #'
 #'
 #'    }
 #'
 simdata <- function(a, d, N, itemtype, sigma = NULL, mu = NULL, guess = 0,
-	upper = 1, nominal = NULL, Theta = NULL)
+	upper = 1, nominal = NULL, Theta = NULL, returnList = FALSE)
 {
-    fn <- function(p, ns) sample(1L:ns, 1, prob = p)
+    fn <- function(p, ns) sample(1L:ns, 1L, prob = p)
 	nfact <- ncol(a)
 	nitems <- nrow(a)
-	K <- rep(0,nitems)
+	K <- rep(0L,nitems)
 	if(length(guess) == 1L) guess <- rep(guess,nitems)
 	if(length(guess) != nitems) stop("Guessing parameter is incorrect")
 	if(length(upper) == 1L) upper <- rep(upper,nitems)
@@ -212,6 +220,7 @@ simdata <- function(a, d, N, itemtype, sigma = NULL, mu = NULL, guess = 0,
     if(is.null(nominal)) nominal <- matrix(NA, nitems, max(K))
 	data <- matrix(0, N, nitems)
     a[is.na(a)] <- 0
+    itemobjects <- vector('list', nitems)
 	for(i in 1L:nitems){
 	    if(itemtype[i] == 'nestlogit'){
 	        par <- na.omit(c(a[i, ],d[i,1], guess[i], upper[i], nominal[i,-1L],d[i,-1L]))
@@ -227,10 +236,15 @@ simdata <- function(a, d, N, itemtype, sigma = NULL, mu = NULL, guess = 0,
         if(any(itemtype[i] == c('gpcm','nominal', 'nestlogit')))
             obj@ncat <- K[i]
         P <- ProbTrace(obj, Theta)
-        data[,i] <- apply(P, 1, fn, ns = ncol(P))
+        data[,i] <- apply(P, 1L, fn, ns = ncol(P))
         if(any(itemtype[i] == c('dich', 'gpcm', 'partcomp'))) data[ ,i] <- data[ ,i] - 1L
+        itemobjects[[i]] <- obj
 	}
 	colnames(data) <- paste("Item_", 1L:nitems, sep="")
-	return(data)
+    if(returnList){
+        return(list(itemobjects=itemobjects, data=data, Theta=Theta))        
+    } else {
+	    return(data)
+    }
 }